// src/app/model/feeds.ts
export class feeds {
    key: string;
    Date: number;
    Name: string;
    Photo: string;
    Score: number;
    UserId: string;
    city: string;
    commentLength: number;
    comments: [];
    description: string;
    groupId: string;
    groupName: string;
    haveILiked: boolean;
    like: [];
    sensitive: boolean;
    video: string;
    images: [];
    location: string;
    latitude: number;
    longitude: number;
}